Refactor of header bar
