+++ src/app-header/menus/ProfileMenu.js
@@ -5,7 +5,7 @@ import HeaderMenu from './HeaderMenu';
 import HeaderMenuItem from './HeaderMenuItem';
 import styles from '../header-bar-styles';
 import addD2Context from '../../component-helpers/addD2Context';
-import getBaseUrlFromD2ApiUrl from '../getBaseUrlFromD2ApiUrl';
+import getBaseUrlFromD2ApiUrl from '../utils/getBaseUrlFromD2ApiUrl';
 
 const getBaseUrl = getBaseUrlFromD2ApiUrl;
