+++ src/app-header/search/sources/settings-app.js
@@ -1,7 +1,7 @@
 import { map } from 'lodash/fp';
 import { Observable } from 'rxjs';
 import { config, getInstance as getD2 } from 'd2/lib/d2';
-import getBaseUrlFromD2ApiUrl from '../../getBaseUrlFromD2ApiUrl';
+import getBaseUrlFromD2ApiUrl from '../../utils/getBaseUrlFromD2ApiUrl';
 import { prepareMenuItems, translate$, translateMenuItemNames } from '../../headerBar.store';
