+++ src/app-header/utils/getBaseUrlFromD2ApiUrl.js
