+++ src/app-header/utils/menu-sources.js
@@ -1,7 +1,7 @@
 import { Observable } from 'rxjs';
 import { getInstance, config } from 'd2/lib/d2';
 import log from 'loglevel';
-import getBaseUrlFromD2ApiUrl from './getBaseUrlFromD2ApiUrl';
+import getBaseUrlFromD2ApiUrl from '../utils/getBaseUrlFromD2ApiUrl';
 
 // Profile menu
 config.i18n.strings.add('settings');
